+++ piker/clearing/__init__.py
@@ -26,12 +26,14 @@ from ._client import (
 from ._ems import (
     open_brokerd_dialog,
 )
+from ._util import OrderDialogs
 
 
 __all__ = [
     'open_ems',
     'OrderClient',
     'open_brokerd_dialog',
+    'OrderDialogs',
 
 ]
